test: add some tests for my application, first test written, they should be refactored in the future
